feat: Add CentOS Stream 10 kernel builder Podman container
This commit introduces a new Podman container configuration to provide a complete
and isolated environment for building the Linux kernel on CentOS Stream 10.

It includes all required build dependencies, enables the CRB and EPEL repositories,
and compiles/installs the `mkbootimg` utility for handling boot images. A
`README.md` is also added with instructions for building and running the container

Signed-off-by: Enric Balletbo i Serra <eballetbo@kernel.org>